+++ src/Helpers/Commands.ts
@@ -3,11 +3,13 @@ export class Commands {
         return 'pac powerpages list';
     }
 
-    public static DownloadPortal(localPath: string, websiteId: string, overwritePortal: string | undefined, includedEntities?: string, excludedEntities?: string) {
+    public static DownloadPortal(localPath: string, websiteId: string, overwritePortal: string | undefined, 
+        includedEntities?: string, excludedEntities?: string, modelVersion?: string) {
         const overwriteText = overwritePortal === 'Yes' ? `-o true` : '';
         const includeEntities = includedEntities ? `-ie "${includedEntities}"`: '';
         const excludeEntities = excludedEntities ? `-xe "${excludedEntities}"`: '';
-        return `pac powerpages download -p "${localPath}" -id ${websiteId} ${overwriteText} ${includeEntities} ${excludeEntities}`;
+        const mVersion = modelVersion ? `-mv "${modelVersion}"`: '';
+        return `pac powerpages download -p "${localPath}" -id ${websiteId} ${overwriteText} ${includeEntities} ${excludeEntities} ${mVersion}`;
     }
 
     public static UploadPortal(localPath: string, deploymentProfile: string | undefined) {